Build / test improvements
1. Got rid of `tool/tsh/common` package. See below.
2. Fixed logger in config test (it was getting reset by the test itself)

The reason we don't need `tool/tsh/common` is because `tsh` is the same
for OSS or Enterprise versions. This is good for two reasons:

1. Enterprise customers don't need to bother telling users to use a
   proprietary binary, they can just get OSS (which will eventually get
   pacakged for OSX/Ubuntu/RHEL) and will be one `apt-get` or `brew install` away

2. Easier for us to package/maintain.
