+++ tool/tsh/tsh_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package common
+package main
 
 import (
 	"fmt"
@@ -97,7 +97,7 @@ func (s *MainTestSuite) TestIdentityRead(c *check.C) {
 	}
 	for _, id := range ids {
 		// test reading:
-		k, cb, err := loadIdentity(fmt.Sprintf("../../../fixtures/certs/identities/%s", id))
+		k, cb, err := loadIdentity(fmt.Sprintf("../../fixtures/certs/identities/%s", id))
 		c.Assert(err, check.IsNil)
 		c.Assert(k, check.NotNil)
 		c.Assert(cb, check.IsNil)
@@ -107,17 +107,17 @@ func (s *MainTestSuite) TestIdentityRead(c *check.C) {
 		c.Assert(err, check.IsNil)
 		c.Assert(am, check.NotNil)
 	}
-	k, _, err := loadIdentity("../../../fixtures/certs/identities/lonekey")
+	k, _, err := loadIdentity("../../fixtures/certs/identities/lonekey")
 	c.Assert(k, check.IsNil)
 	c.Assert(err, check.NotNil)
 
 	// lets read an indentity which includes a CA cert
-	k, hostAuthCallback, err := loadIdentity("../../../fixtures/certs/identities/key-cert-ca.pem")
+	k, hostAuthCallback, err := loadIdentity("../../fixtures/certs/identities/key-cert-ca.pem")
 	c.Assert(err, check.IsNil)
 	c.Assert(k, check.NotNil)
 	c.Assert(hostAuthCallback, check.NotNil)
 	// prepare the cluster CA separately
-	certBytes, err := ioutil.ReadFile("../../../fixtures/certs/identities/ca.pem")
+	certBytes, err := ioutil.ReadFile("../../fixtures/certs/identities/ca.pem")
 	c.Assert(err, check.IsNil)
 	_, hosts, cert, _, _, err := ssh.ParseKnownHosts(certBytes)
 	c.Assert(err, check.IsNil)
